requirements.txt
openai
tiktoken
chromadb
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
sentence-transformers
langchain-chroma
flask
python-dotenv
gitpython
numpy
-e .